Fabryka - done

+++ include/types.hpp
@@ -2,10 +2,10 @@
 
 #include <functional>
 
-using ElementID = unsigned long;
+using ElementID = unsigned long long;
 
 using ProbabilityGenerator = std::function<double()>;
 
-using TimeOffset = unsigned long;
+using TimeOffset = unsigned long long;
 
-using Time = unsigned long;
+using Time = unsigned long long;
